Модуль подсчета транзакций и юнит тест

+++ app/src/main/java/com/example/beardie/currencyholder/data/model/ExchangeCurrency.kt
@@ -6,5 +6,5 @@ data class ExchangeCurrency (
     var id: String = UUID.randomUUID().toString(),
     var fromCurrency: FinanceCurrency,
     var toCurrency: FinanceCurrency,
-    var coef: Float
+    var coef: Double
 )
